requirements.txt
functions-framework==3.*
pandas
google-cloud-bigquery
google-cloud-storage
pandas_gbq